requirements.txt
aiofiles==0.6.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.7.1
APScheduler==3.6.3
async-lru==1.0.2
async-timeout==4.0.2
attrs==22.2.0
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==2.1.1
coverage==6.2
dataclass-factory==2.16
exceptiongroup==1.1.2
frozenlist==1.3.3
google-api-core==1.29.0
google-api-python-client==2.7.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.21.0
httpx==0.24.1
idna==3.4
iniconfig==1.1.1
multidict==6.0.4
mutagen==1.45.1
numpy==1.24.1
oauthlib==3.2.2
packaging==22.0
pluggy==1.0.0
protobuf==4.21.12
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.24.1
pyparsing==3.0.9
Pyrogram==1.2.20
PySocks==1.7.1
pytest==6.2.5
pytest-asyncio==0.16.0
pytest-mock==3.6.1
python-decouple==3.3
python-magic==0.4.24
python-telegram-bot==20.3
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
shazamio==0.0.5
six==1.16.0
sniffio==1.3.0
tgintegration==1.2.0
toml==0.10.2
tornado==6.2
typing-extensions==3.10.0.2
tzdata==2022.7
tzlocal==4.2
uritemplate==3.0.1
urllib3==1.26.13
yarl==1.8.2
youtube-dl==2021.6.6
